migration in preparation for rev1

+++ LuckParser/Models/ParseModels/Boons/BoonDistribution.cs
@@ -15,13 +15,13 @@ namespace LuckParser.Models.ParseModels
         }
     }
 
-    public class BoonDistribution : Dictionary<int, Dictionary<ushort, OverAndValue>>
+    public class BoonDistribution : Dictionary<long, Dictionary<ushort, OverAndValue>>
     {
         public BoonDistribution() : base()
         {
         }
 
-        public long getUptime(int boonid)
+        public long getUptime(long boonid)
         {
             if (!ContainsKey(boonid))
             {
@@ -30,7 +30,7 @@ namespace LuckParser.Models.ParseModels
             return this[boonid].Sum(x => x.Value.value);
         }
 
-        public long getGeneration(int boonid, ushort src)
+        public long getGeneration(long boonid, ushort src)
         {
             if (!ContainsKey(boonid) || src == 0)
             {
@@ -39,7 +39,7 @@ namespace LuckParser.Models.ParseModels
             return this[boonid].Where( x => src == x.Key).Sum(x => x.Value.value);
         }
 
-        public long getOverstack(int boonid, ushort src)
+        public long getOverstack(long boonid, ushort src)
         {
             if (!ContainsKey(boonid) || src == 0)
             {
